+++ lin_actuator.c
@@ -1,5 +1,23 @@
 #include <xc.h>
+#include<stdint.h>
+#include "mcc_generated_files/i2c1.h"
+#include "mcc_generated_files/mcc.h"
 #include "lin_actuator.h"
+
+static lin_actuator_states moving = stopped;
+
+void set_DACs(){
+  uint8_t data_buf[2]; 
+  data_buf[0] = ((HIGH_DAC_VAL & 0xf0) >> 4); // shift the 4 MSB to the 4 LSB
+  data_buf[1] = ((HIGH_DAC_VAL & 0x0f) << 4); // mask the 4MSB then shift the 4 LSB to the 4MSB
+  i2c1_writeNBytes(HIGH_DAC_ADDRESS, data_buf, 2);
+  data_buf[0] = ((LOW_DAC_VAL & 0xf0) >> 4); // shift the 4 MSB to the 4 LSB
+  data_buf[1] = ((LOW_DAC_VAL) << 4); // mask the 4MSB then shift the 4 LSB to the 4MSB
+  i2c1_writeNBytes(LOW_DAC_ADDRESS, data_buf, 2); 
+  if(i2c1_getLastError() == I2C1_FAIL_TIMEOUT){
+        RED_LED_ON;
+  }
+}
 
  void lin_actuator_init() {
     TRISB1 = 0;     //set MOTOR_FWD as output
