+++ lin_actuator.h
@@ -1,6 +1,6 @@
 /* 
  * File:   LINAC.h
- * Author: zx100
+ * Author: Zachariah Mears
  *
  * Created on February 10, 2019, 12:25 AM
  */
@@ -8,20 +8,30 @@
 #ifndef LINAC_H
 #define	LINAC_H
 
+#define RED_LED_ON (LATC5 = 0)
+#define RED_LED_OFF (LATC5 = 1)
+#define WHITE_LED_ON (LATC6 = 0)
+#define WHITE_LED_OFF (LATC6 = 1)
+#define BLUE_LED_ON (LATC7 = 0)
+#define BLUE_LED_OFF (LATC7 = 1)
+#define HIGH_DAC_ADDRESS 0x4d
+#define LOW_DAC_ADDRESS 0x4c
+#define HIGH_DAC_VAL 51 //from 0-255, multiply  desired voltage by 51
+#define LOW_DAC_VAL 5
+
 typedef enum
 {
-    nominal =  0,
-    open = 1,
-    closed = 2,
-    opening = 3,
-    closing = 4,
-    stopped_between_states =  5,
-    in_motion =  6,
-    stopped = 7
+    nominal,
+    open,
+    closed,
+    opening,
+    closing,
+    stopped_between_states,
+    in_motion,
+    stopped
 } lin_actuator_states;
 
-
-lin_actuator_states moving = stopped;
+void set_DACs();
 
 void lin_actuator_init(void);
